msk-iam-python-provisioning/bastion-resources/pysamples/iam_requirements.txt
Faker
confluent-kafka==2.3.0
aws-msk-iam-sasl-signer-python>=1.0.0,<2.0.0
